Update collision detection method and remove level data
Modified the `detectCollisions` method in the `ICollision` class to return a Boolean value for better handling of collision logic. Adjusted how the separation vector is calculated for circle type collisions. Removed the `level.json` file as it was no longer needed and refactored the level data into a new `level1.json` file.