Update movement calculations to factor in delta time
The moveByInputVector function in both the main source file and the ITransform class has been updated to include delta time in the calculations, which allows for smoother and more accurate movements. The same change has been implemented in the translate function of the ITransform class. Furthermore, a bug resulting in misplaced respawn position was fixed.